GEODE-3031: Extracting startLocator and startServer from LauncherLifecycleCommands
