+++ geode-assembly/src/test/java/org/apache/geode/management/internal/cli/commands/GfshCommandIntegrationTest.java
@@ -12,8 +12,8 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.management.internal.cli.commands;
 
+package org.apache.geode.management.internal.cli.commands;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
@@ -30,13 +30,10 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.jar.Attributes;
-import java.util.jar.Attributes.Name;
 import java.util.jar.JarFile;
 import java.util.jar.Manifest;
 
 import org.apache.commons.io.FileUtils;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -46,52 +43,46 @@ import org.junit.rules.TestName;
 import org.apache.geode.internal.util.IOUtils;
 import org.apache.geode.test.junit.categories.IntegrationTest;
 
-/**
- * The LauncherLifecycleCommandsIntegrationTest class is a test suite of test cases testing the
- * contract and functionality of the lifecycle launcher GemFire shell (Gfsh) commands.
- *
- * @see org.apache.geode.management.internal.cli.commands.LauncherLifecycleCommands
- * @see org.junit.Assert
- * @see org.junit.Test
- * @since GemFire 7.0
- */
 @Category(IntegrationTest.class)
-public class LauncherLifecycleCommandsIntegrationTest {
-
-  private LauncherLifecycleCommands launcherCommands;
-
+public class GfshCommandIntegrationTest {
   @Rule
   public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
   @Rule
   public TestName testName = new TestName();
 
-  @Before
-  public void setup() {
-    launcherCommands = new LauncherLifecycleCommands();
-  }
-
-  @After
-  public void tearDown() {
-    launcherCommands = null;
+  @Test
+  public void workingDirDefaultsToMemberName() {
+    StartServerCommand startServer = new StartServerCommand();
+    String workingDir = StartMemberUtils.resolveWorkingDir(null, "server1");
+    assertThat(new File(workingDir)).exists();
+    assertThat(workingDir).endsWith("server1");
   }
 
   @Test
-  public void testGemFireCoreClasspath() throws IOException {
-    File coreDependenciesJar = new File(LauncherLifecycleCommands.CORE_DEPENDENCIES_JAR_PATHNAME);
+  public void workingDirGetsCreatedIfNecessary() throws Exception {
+    File workingDir = temporaryFolder.newFolder("foo");
+    FileUtils.deleteQuietly(workingDir);
+    String workingDirString = workingDir.getAbsolutePath();
 
-    assertNotNull(coreDependenciesJar);
-    assertTrue(coreDependenciesJar + " is not a file", coreDependenciesJar.isFile());
+    StartServerCommand startServer = new StartServerCommand();
 
-    Collection<String> expectedJarDependencies =
-        Arrays.asList("antlr", "commons-io", "commons-lang", "commons-logging", "geode",
-            "jackson-annotations", "jackson-core", "jackson-databind", "jansi", "jline", "snappy",
-            "spring-core", "spring-shell", "jetty-server", "jetty-servlet", "jetty-webapp",
-            "jetty-util", "jetty-http", "servlet-api", "jetty-io", "jetty-security", "jetty-xml"
+    String resolvedWorkingDir = StartMemberUtils.resolveWorkingDir(workingDirString, "server1");
+    assertThat(new File(resolvedWorkingDir)).exists();
+    assertThat(workingDirString).endsWith("foo");
+  }
 
-        );
+  @Test
+  public void testWorkingDirWithRelativePath() throws Exception {
+    Path relativePath = Paths.get("some").resolve("relative").resolve("path");
+    assertThat(relativePath.isAbsolute()).isFalse();
 
-    assertJarFileManifestClassPath(coreDependenciesJar, expectedJarDependencies);
+    StartServerCommand startServer = new StartServerCommand();
+
+    String resolvedWorkingDir =
+        StartMemberUtils.resolveWorkingDir(relativePath.toString(), "server1");
+
+    assertThat(resolvedWorkingDir).isEqualTo(relativePath.toAbsolutePath().toString());
   }
 
   @Test
@@ -107,21 +98,22 @@ public class LauncherLifecycleCommandsIntegrationTest {
     pidFile.deleteOnExit();
     writePid(pidFile, expectedPid);
 
-    final int actualPid = getLauncherLifecycleCommands().readPid(pidFile);
+    final int actualPid = StartMemberUtils.readPid(pidFile);
 
     assertEquals(expectedPid, actualPid);
   }
 
-  private LauncherLifecycleCommands getLauncherLifecycleCommands() {
-    return launcherCommands;
-  }
-
-  private void writePid(final File pidFile, final int pid) throws IOException {
-    final FileWriter fileWriter = new FileWriter(pidFile, false);
-    fileWriter.write(String.valueOf(pid));
-    fileWriter.write("\n");
-    fileWriter.flush();
-    IOUtils.close(fileWriter);
+  @Test
+  public void testGemFireCoreClasspath() throws IOException {
+    File coreDependenciesJar = new File(StartMemberUtils.CORE_DEPENDENCIES_JAR_PATHNAME);
+    assertNotNull(coreDependenciesJar);
+    assertTrue(coreDependenciesJar + " is not a file", coreDependenciesJar.isFile());
+    Collection<String> expectedJarDependencies =
+        Arrays.asList("antlr", "commons-io", "commons-lang", "commons-logging", "geode",
+            "jackson-annotations", "jackson-core", "jackson-databind", "jansi", "jline", "snappy",
+            "spring-core", "spring-shell", "jetty-server", "jetty-servlet", "jetty-webapp",
+            "jetty-util", "jetty-http", "servlet-api", "jetty-io", "jetty-security", "jetty-xml");
+    assertJarFileManifestClassPath(coreDependenciesJar, expectedJarDependencies);
   }
 
   private void assertJarFileManifestClassPath(final File dependenciesJar,
@@ -134,9 +126,9 @@ public class LauncherLifecycleCommandsIntegrationTest {
     Attributes attributes = manifest.getMainAttributes();
 
     assertNotNull(attributes);
-    assertTrue(attributes.containsKey(Name.CLASS_PATH));
+    assertTrue(attributes.containsKey(Attributes.Name.CLASS_PATH));
 
-    String[] actualJarDependencies = attributes.getValue(Name.CLASS_PATH).split(" ");
+    String[] actualJarDependencies = attributes.getValue(Attributes.Name.CLASS_PATH).split(" ");
 
     assertNotNull(actualJarDependencies);
     assertTrue(
@@ -171,44 +163,15 @@ public class LauncherLifecycleCommandsIntegrationTest {
         String.format(
             "GemFire dependencies JAR file (%1$s) does not contain the expected dependencies (%2$s) in the Manifest Class-Path attribute (%3$s)!",
             dependenciesJar, missingExpectedJarDependenciesList,
-            attributes.getValue(Name.CLASS_PATH)),
+            attributes.getValue(Attributes.Name.CLASS_PATH)),
         missingExpectedJarDependenciesList.isEmpty());
   }
 
-  @Test
-  public void workingDirDefaultsToMemberName() {
-    LauncherLifecycleCommands launcherLifecycleCommands = new LauncherLifecycleCommands();
-
-    String workingDir = launcherLifecycleCommands.resolveWorkingDir(null, "server1");
-    assertThat(new File(workingDir)).exists();
-    assertThat(workingDir).endsWith("server1");
-  }
-
-
-  @Test
-  public void workingDirGetsCreatedIfNecessary() throws Exception {
-    File workingDir = temporaryFolder.newFolder("foo");
-    FileUtils.deleteQuietly(workingDir);
-    String workingDirString = workingDir.getAbsolutePath();
-
-    LauncherLifecycleCommands launcherLifecycleCommands = new LauncherLifecycleCommands();
-
-    String resolvedWorkingDir =
-        launcherLifecycleCommands.resolveWorkingDir(workingDirString, "server1");
-    assertThat(new File(resolvedWorkingDir)).exists();
-    assertThat(workingDirString).endsWith("foo");
-  }
-
-  @Test
-  public void testWorkingDirWithRelativePath() throws Exception {
-    Path relativePath = Paths.get("some").resolve("relative").resolve("path");
-    assertThat(relativePath.isAbsolute()).isFalse();
-
-    LauncherLifecycleCommands launcherLifecycleCommands = new LauncherLifecycleCommands();
-
-    String resolvedWorkingDir =
-        launcherLifecycleCommands.resolveWorkingDir(relativePath.toString(), "server1");
-
-    assertThat(resolvedWorkingDir).isEqualTo(relativePath.toAbsolutePath().toString());
+  private void writePid(final File pidFile, final int pid) throws IOException {
+    final FileWriter fileWriter = new FileWriter(pidFile, false);
+    fileWriter.write(String.valueOf(pid));
+    fileWriter.write("\n");
+    fileWriter.flush();
+    IOUtils.close(fileWriter);
   }
 }
